#232: hold every setting in one module
django_logic/conf.py now owns every DJANGO_LOGIC key: one reader per
key, one number validator, one bool validator, one place per default.
django_logic/background/settings.py is gone; it already imported the
private _conf reader and called the core gate, so the split was crossed
in both directions.

The two boot gates keep their split so a synchronous-only install does
not fail on background keys: the core app validates the core keys; the
background app's gate (now in background/apps.py, with the two
pull-mode deployment checks) validates the background keys on top.
