+++ django_logic/conf.py
@@ -1,14 +1,14 @@
-"""Core ``DJANGO_LOGIC`` settings — readers and boot-time validation.
+"""Every ``DJANGO_LOGIC`` settings reader, and boot-time validation.
 
-These knobs are consumed by the core engine (state locks, transition
-unlock semantics) independently of the optional
-``django_logic.background`` app, so their validation cannot live only in
-the background app's ready hook: a sync-only install that registers just
-``django_logic`` must fail fast on misconfiguration too.
-``DjangoLogicConfig.ready`` calls :func:`validate_core_settings`;
-``django_logic.background.settings.validate_on_ready`` calls the same
-function as part of its full safety gate (both paths are idempotent —
-pure reads, no state).
+One module owns every key: one reader per key, one number validator,
+one bool validator, and one place where each default is written down.
+
+Validation runs from two boot gates. ``DjangoLogicConfig.ready`` calls
+:func:`validate_core_settings` — the keys the core engine reads with or
+without the background app installed. The background app's ready hook
+(``django_logic.background.apps``) validates the background keys and
+the pull-mode deployment requirements on top. Both paths are idempotent
+— pure reads, no state.
 """
 import math
 
@@ -16,6 +16,10 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 
 LOCK_TIMEOUT_DEFAULT = 7200
+
+EXECUTION_SYNC = 'sync'
+EXECUTION_PULL = 'pull'
+_VALID_EXECUTION_MODES = frozenset({EXECUTION_SYNC, EXECUTION_PULL})
 
 
 def _conf() -> dict:
@@ -64,6 +68,145 @@ def legacy_exception_base():
     return _conf().get('LEGACY_EXCEPTION_BASE') or None
 
 
+def background_execution() -> str:
+    """Return the configured execution mode.
+
+    Defaults to ``'pull'`` — workers claim committed rows from the
+    database (run them with ``manage.py dl_worker``). ``'sync'`` runs the
+    worker inline in the same process and exists for tests, CI,
+    management commands, and the shell.
+    """
+    configured = _conf().get('BACKGROUND_EXECUTION')
+    if configured is None:
+        return EXECUTION_PULL
+    if configured == 'celery':
+        raise ImproperlyConfigured(
+            "DJANGO_LOGIC['BACKGROUND_EXECUTION']='celery' was removed: "
+            "workers now claim committed rows from the database, so no "
+            "broker carries them. Set 'pull' and run one "
+            "`manage.py dl_worker --queues <names>` process per queue "
+            "group (the worker loop also runs the safety nets, so no "
+            "beat schedule is needed). Drain the old broker queues once "
+            "before switching. See the README's background section."
+        )
+    if configured not in _VALID_EXECUTION_MODES:
+        raise ImproperlyConfigured(
+            f"DJANGO_LOGIC['BACKGROUND_EXECUTION'] must be one of "
+            f"{sorted(_VALID_EXECUTION_MODES)}; got {configured!r}."
+        )
+    return configured
+
+
+def default_queue() -> str:
+    """Queue used by background transitions that don't declare ``queue=``.
+
+    Per-transition ``queue=`` overrides this; use it to route work to
+    dedicated workers (e.g. ``critical`` / ``slow``) and manage
+    performance per queue.
+    """
+    queue = _conf().get('DEFAULT_QUEUE', 'django_logic')
+    if not queue or not isinstance(queue, str):
+        raise ImproperlyConfigured(
+            "DJANGO_LOGIC['DEFAULT_QUEUE'] must be a non-empty string."
+        )
+    return queue
+
+
+def _validated_number(
+    key: str,
+    default,
+    *,
+    minimum,
+    integral: bool = False,
+):
+    """Read ``DJANGO_LOGIC[key]`` and validate it is a sane number.
+
+    Raises ``ImproperlyConfigured`` naming the setting and the offending
+    value. ``bool`` is rejected explicitly (it subclasses ``int``, so
+    ``True`` would otherwise pass as ``1``); non-finite floats (``nan``,
+    ``inf``) are rejected; ``integral=True`` additionally rejects
+    non-integral floats and returns an ``int``. ``minimum`` is inclusive.
+    """
+    value = _conf().get(key, default)
+    if isinstance(value, bool) or not isinstance(value, (int, float)):
+        raise ImproperlyConfigured(
+            f"DJANGO_LOGIC[{key!r}] must be a number, got {value!r}."
+        )
+    if isinstance(value, float) and not math.isfinite(value):
+        raise ImproperlyConfigured(
+            f"DJANGO_LOGIC[{key!r}] must be a finite number, got {value!r}."
+        )
+    if integral and isinstance(value, float) and not value.is_integer():
+        raise ImproperlyConfigured(
+            f"DJANGO_LOGIC[{key!r}] must be a whole number, got {value!r}."
+        )
+    if value < minimum:
+        raise ImproperlyConfigured(
+            f"DJANGO_LOGIC[{key!r}] must be >= {minimum}, got {value!r}."
+        )
+    if integral:
+        return int(value)
+    return value
+
+
+def max_errors() -> int:
+    """Attempts before a background transition is finalized as failed.
+    Must be a whole number >= 1 (0 would finalize before the first
+    attempt ever ran)."""
+    return _validated_number(
+        'TRANSITION_MESSAGE_MAX_ERRORS', 5, minimum=1, integral=True)
+
+
+def retry_minutes():
+    """Minutes a failed row stays unclaimable after ``last_error_dt``.
+    The pull claim's ``WHERE`` clause is the retry rule: nothing
+    re-dispatches a row; it simply becomes claimable again after this
+    wait. Must be >= 0; zero means "claimable immediately" and is used
+    by tests to retry without back-dating rows."""
+    return _validated_number(
+        'TRANSITION_MESSAGE_RETRY_MINUTES', 2, minimum=0)
+
+
+def cleanup_days():
+    """Age (days) before completed rows are deleted by the periodic
+    cleanup. Must be >= 0. Zero deletes every completed row on the next
+    cleanup tick — that erases the audit trail, so it is test-only."""
+    return _validated_number(
+        'TRANSITION_MESSAGE_CLEANUP_DAYS', 7, minimum=0)
+
+
+def validate_bool(key: str) -> None:
+    """Reject a non-bool on a setting that gates behaviour.
+
+    Truthiness coercion is unsafe for these: the strings 'false'/'no'/'0'
+    are all truthy, so a value meant to disable a feature enabled it.
+    """
+    value = _conf().get(key, False)
+    if not isinstance(value, bool):
+        raise ImproperlyConfigured(
+            f"DJANGO_LOGIC[{key!r}] must be a bool (True or False), got "
+            f"{value!r}. Strings are not accepted — 'false' would otherwise "
+            f"read as truthy and enable the very behaviour it names."
+        )
+
+
+def strict_kwargs_serialization() -> bool:
+    """When True, enqueue kwargs serialization raises on silently-droppable
+    caller kwargs (``request``) instead of logging a warning.
+
+    Default False: generic API layers commonly pass ``request`` to every
+    transition uniformly, so raising by default would break them. Enable
+    once call sites are clean to turn the drop into a hard contract.
+
+    Only a literal ``True`` enables it. It used to be
+    ``bool(...)``-coerced, so any non-empty string switched strict mode ON —
+    reading ``DL_STRICT=false`` from an env var made enqueue start raising.
+    Mirrors :func:`defer_unlock_until_commit`; boot validation rejects
+    non-bools and this reader stays safe where that has not run.
+    """
+    return _conf().get('STRICT_KWARGS_SERIALIZATION', False) is True
+
+
 def validate_core_settings() -> None:
     """Fail fast on misconfigured core knobs (``ImproperlyConfigured``
     naming the setting), from every install shape."""
@@ -81,17 +224,11 @@ def validate_core_settings() -> None:
         )
     # Both strict flags are read with `is True`, so truthy garbage disables
     # them rather than enabling — silent in the UNSAFE direction. Validated
-    # here (not in background/settings) because STRICT_HOOK_SIGNATURES is a
-    # core setting read from process.py at bind time, with no background app
-    # involved, so a sync-only install must be covered too.
+    # here because STRICT_HOOK_SIGNATURES is a core setting read from
+    # process.py at bind time, with no background app involved, so a
+    # sync-only install must be covered too.
     for key in ('DEFER_UNLOCK_UNTIL_COMMIT', 'STRICT_HOOK_SIGNATURES'):
-        value = _conf().get(key, False)
-        if not isinstance(value, bool):
-            raise ImproperlyConfigured(
-                f"DJANGO_LOGIC[{key!r}] must be a bool (True or False), got "
-                f"{value!r}. Strings are not accepted — 'false' would "
-                f"otherwise read as truthy."
-            )
+        validate_bool(key)
     # Type check only — resolving the dotted path imports consumer code,
     # which does not belong in a pure-read validator. The import happens in
     # install_legacy_exception_base(), where failures are equally loud and
